fix: accounting for multiple catalogs in catalog reporting
